+++ src/Tgraph/Grid.hs
@@ -17,20 +17,30 @@ It is used to quickly identify touching vertices when forcing.
 module Tgraph.Grid
 (
     insertGridCheck
+  , createGridCheck
   , createGrid
+  , createPointGrid
+  , emptyGrid
   , Grid
+  , touching
   
 ) where
 
 import Diagrams.Prelude
-import Tgraph.Prelude ( touching )
 
+
+import Data.List ( find )
 import Data.IntMap.Strict (IntMap)
 import qualified Data.IntMap.Strict as IMap (alter,insert,empty,lookup)
 
--- | A Grid maps from Int x and y coordinates to a list of points in the unit square
--- with bottom left at the coordinates.
-newtype Grid = Grid{ gridmap::IntMap (IntMap [P2 Double])} deriving (Show)
+-- | A (Grid a) maps from Int x and y coordinates to a list of valued points in the unit square
+-- with bottom left corner given by the coordinates, each valued point is a pair with a value and point.
+-- For a Grid with just points (and no associated values), use Grid ().
+newtype Grid a = Grid{ gridmap::IntMap (IntMap [(a,P2 Double)])} deriving (Show)
+
+-- | An empty grid.
+emptyGrid :: Grid a
+emptyGrid = Grid IMap.empty
 
 -- | get the Int x and Int y coordinates of a 2D point (using floor).
 gridCoords :: P2 Double -> (Int,Int)
@@ -38,36 +48,37 @@ gridCoords p = (floor x, floor y)
                where x = p ^. _x 
                      y = p ^. _y 
 
--- | insert a new point in a grid, but check the 9 neighbouring cells for any
--- points touching the new point.  If a touching point is found, Nothing is returned,
--- otherwise Just gd' is returned where gd' is an updated grid with the new point.
-insertGridCheck :: P2 Double -> Grid -> Maybe Grid
-insertGridCheck p gd =
-    case checkNear p gd of
-        True -> Nothing
-        False -> Just $ insertGrid p gd
+-- | insert a new (value,point) in a grid, but check the 9 neighbouring cells for any
+-- points touching the new point.  If a touching point is found, a (Left v) is returned,
+-- where v was associated with the (first found) touching point.
+-- Otherwise Right gd' is returned where gd' is an updated grid with the new (valued) point.
+insertGridCheck :: (a, P2 Double) -> Grid a -> Either a (Grid a)
+insertGridCheck ap@(_,p) gd =
+    case checkGridNear p gd of
+        Just b  -> Left b
+        Nothing -> Right $ insertGrid ap gd
 
- -- | insert a new point in a grid without any checks.
- -- This is used to initialise a grid with points known not to be touching.       
-insertGrid :: P2 Double -> Grid -> Grid
-insertGrid p gd = Grid $ IMap.alter column n (gridmap gd)
-   where (n,m) = gridCoords p
-         column Nothing = Just $ IMap.insert n [p] IMap.empty
+ -- | insert a new (valued) point in a grid without any checks.
+ -- This is used to initialise a grid with (valued) where the points are known not to be touching.       
+insertGrid :: (a, P2 Double) -> Grid a -> Grid a
+insertGrid ap gd = Grid $ IMap.alter column n (gridmap gd)
+   where (n,m) = gridCoords (snd ap)
+         column Nothing = Just $ IMap.insert n [ap] IMap.empty
          column (Just imp) = Just $ IMap.alter ht m imp
-         ht Nothing = Just [p]
-         ht (Just ps) = Just (p:ps)
+         ht Nothing = Just [ap]
+         ht (Just aps) = Just (ap:aps)
 
--- | get the list of points in a grid cell (with Int coords x y)         
-fromGrid :: Grid -> Int -> Int -> [P2 Double]
+-- | get the list of (valued) points in a grid cell (with Int coords x y)         
+fromGrid :: Grid a -> Int -> Int -> [(a,P2 Double)]
 fromGrid gd n = 
     case IMap.lookup n (gridmap gd) of
       Nothing ->  (\_ -> [])
       Just imp -> (\m -> case IMap.lookup m imp of
                           Nothing -> []
-                          Just ps -> ps)
+                          Just aps -> aps)
 
--- | get the list of points from 9 grid cells (aound the one with Int coords x y)         
-fromGridNear :: Grid -> Int -> Int -> [P2 Double]
+-- | get the list of (valued) points from 9 grid cells (aound the one with Int coords x y)         
+fromGridNear :: Grid a -> Int -> Int -> [(a,P2 Double)]
 fromGridNear gd n m =  
     let g0 = fromGrid gd n
         g1 = fromGrid gd (n-1)
@@ -78,12 +89,40 @@ fromGridNear gd n m =
               ]
 
 -- | check a point for any touching locations in the grid
--- (checking with points in the 9 grid cells near the point).        
-checkNear :: P2 Double -> Grid -> Bool             
-checkNear p gd = any (touching p) (fromGridNear gd n m) where
+-- (checking with points in the 9 grid cells near the point).
+-- This will return Nothing if there are no touching locations
+-- and Just v otherwise where v is the value paired with the first touching location found.       
+checkGridNear :: P2 Double -> Grid a -> Maybe a             
+checkGridNear p gd = fmap fst $ find touchingItem (fromGridNear gd n m) where
                  (n,m) = gridCoords p
+                 touchingItem (_,p') = touching p p'
 
- -- | insert points into a new grid without any checks.
- -- This is used to initialise a grid with points known not to be touching.       
-createGrid :: [P2 Double] -> Grid
-createGrid ps = foldl' (flip insertGrid) (Grid IMap.empty) ps
+{-|touching checks if two points are considered close.
+Close means the square of the distance between them is less than a certain number (currently 0.25) so they cannot be
+vertex locations for 2 different vertices in a VPatch using unit scale for short edges.
+-}
+touching :: P2 Double -> P2 Double -> Bool
+touching p p1 = quadrance (p .-. p1) < 0.25 -- quadrance is square of length of a vector
+-- quadrance 0.1 represents a distance of about 0.316 units (= sqrt 0.1)
+-- quadrance 0.25 represents a distance of 0.5 units
+
+                 
+-- | insert (valued) points into a new grid without any checks.
+-- This is used to initialise a grid with (valued) points known not to be touching.       
+createGrid :: [(a,P2 Double)] -> Grid a
+createGrid aps = foldl' (flip insertGrid) (Grid IMap.empty) aps
+
+-- | insert points into a new grid without any checks.
+-- This uses () for values.      
+createPointGrid :: [P2 Double] -> Grid ()
+createPointGrid = createGrid . map ((),) 
+
+-- | insert (valued) points into a new grid with checks.
+-- This will return Left v if there is a touching location with value v,
+-- and Just gd otherwise where gd is a grid containiing the (valued) points.
+createGridCheck :: [(a, P2 Double)] -> Either a (Grid a)
+createGridCheck aps = insertAll aps (Grid IMap.empty) where
+  insertAll [] gd = Right gd
+  insertAll (ap:more) gd = 
+    do gd1 <- insertGridCheck ap gd
+       insertAll more gd1